revert changes in Append & DeleteTo
